+++ projects/ngx-xchange-ui/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.1.1 28/01/2025
+* (change) - list item selector component
+
 # 0.1.0 27/01/2025
 * (add) - button 500 from new iu ux design
 * (add) - button 440 from new IU ux